# docs/api/openapi-daily-bias.yaml
openapi: 3.0.0
info:
  title: Daily Bias Analysis API
  version: 1.0.0
  description: |
    AI-powered 6-step daily bias analysis for trading instruments
    
    ## Overview
    
    The Daily Bias Analysis API provides comprehensive market analysis through 6 sequential steps:
    1. **Security Analysis** - Volatility and risk assessment
    2. **Macro Analysis** - Economic context and events
    3. **Institutional Flux** - Volume and order flow analysis
    4. **Mag 7 Leaders** - Correlation with tech leaders
    5. **Technical Structure** - Support, resistance, and trend analysis
    6. **Synthesis** - Final bias aggregation
    
    ## Authentication
    
    All endpoints require a valid JWT token in the `Authorization` header.
    
    ## Rate Limiting
    
    - 10 requests per second (burst)
    - 1 analysis request per user per day (unless admin)
    - Cache TTL: 5 minutes per instrument
    
    ## Performance Targets
    
    - Individual steps: < 3s (p95)
    - Full analysis: < 10s (p95)
    - Cache hit: < 100ms
    
  contact:
    name: API Support
    url: https://tradingjournal.app/support
  
  license:
    name: Proprietary
    url: https://tradingjournal.app/license
servers:
  - url: https://api.tradingjournal.app
    description: Production
  - url: https://staging-api.tradingjournal.app
    description: Staging
  - url: http://localhost:3000
    description: Development

# ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token required for all endpoints
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            code: "INVALID_INPUT"
            timestamp: "2026-01-17T14:30:00Z"
...
  schemas:
...
    RiskLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    SentimentLevel:
      type: string
      enum: [VERY_BEARISH, BEARISH, NEUTRAL, BULLISH, VERY_BULLISH]
    TrendDirection:
      type: string
      enum: [UP, DOWN, INDETERMINATE]
...
    Mag7Symbol:
      type: string
      enum: [AAPL, MSFT, GOOGL, AMZN, META, NVDA, TSLA]
    
    # Shared objects
    ErrorResponse:
      type: object
      required: [error, code, timestamp]
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          enum: [INVALID_INPUT, RATE_LIMIT, SERVICE_UNAVAILABLE, INTERNAL_ERROR]
        details:
          type: object
          description: Additional error context
        timestamp:
          type: string
          format: date-time
    
    # Step 1: Security Analysis
    VolatilityFactor:
      type: object
      required: [factor, impact]
      properties:
        factor:
          type: string
          example: "Fed announcements expected"
        impact:
          type: string
          enum: [LOW, MEDIUM, HIGH]
    RiskItem:
      type: object
      required: [risk, probability, impact]
      properties:
        risk:
          type: string
          example: "Gap risk"
        probability:
          type: number
          minimum: 0
          maximum: 1
          example: 0.35
        impact:
          type: number
          minimum: 0
          maximum: 1
          example: 0.45
    SecurityAnalysis:
      type: object
      required: [volatilityIndex, riskLevel, securityScore, analysis, timestamp, instrument]
      properties:
        volatilityIndex:
          type: number
          minimum: 0
          maximum: 100
          description: Volatility index (0-100)
        riskLevel:
          $ref: '#/components/schemas/RiskLevel'
        securityScore:
          type: number
          minimum: 0
          maximum: 10
          description: Composite security score
        analysis:
          type: object
          required: [summary, risks, recommendations]
          properties:
            summary:
              type: string
              maxLength: 500
            volatilityFactors:
              type: array
              items:
                $ref: '#/components/schemas/VolatilityFactor'
            risks:
              type: array
              items:
                $ref: '#/components/schemas/RiskItem'
            recommendations:
              type: array
              items:
                type: string
        timestamp:
          type: string
          format: date-time
        instrument:
          type: string
    
    # Step 2: Macro Analysis
...
    Mag7Correlation:
      type: object
      required: [symbol, correlation, trend, strength]
      properties:
        symbol:
          $ref: '#/components/schemas/Mag7Symbol'
        correlation:
          type: number
          minimum: -1
          maximum: 1
        trend:
          $ref: '#/components/schemas/TrendDirection'
        performancePercent:
          type: number
        strength:
          type: number
          minimum: 0
          maximum: 1
    Mag7Analysis:
      type: object
      required: [correlations, leaderScore, sentiment, timestamp, instrument]
      properties:
        correlations:
          type: array
          items:
            $ref: '#/components/schemas/Mag7Correlation'
        leaderScore:
          type: number
          minimum: 0
          maximum: 10
        sentiment:
          $ref: '#/components/schemas/SentimentLevel'
        analysis:
          type: object
          properties:
            summary:
              type: string
              maxLength: 500
            leaderDynamics:
              type: string
            groupSentiment:
              type: array
              items:
                type: object
                properties:
                  category:
                    type: string
                  sentiment:
                    $ref: '#/components/schemas/SentimentLevel'
        timestamp:
          type: string
          format: date-time
        instrument:
          type: string
    
    # Step 5: Technical Structure
...
paths:
...
  /api/daily-bias/security:
    post:
      summary: Security Analysis (Step 1 only)
      operationId: getSecurityAnalysis
      security:
        - BearerAuth: []
      tags:
        - Analysis Steps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instrument]
              properties:
                instrument:
                  type: string
                  example: "NQ1"
      responses:
        '200':
          description: Security analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityAnalysis'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/daily-bias/mag7:
    post:
      summary: Mag 7 Leaders (Step 4 only)
      operationId: getMag7Analysis
      security:
        - BearerAuth: []
      tags:
        - Analysis Steps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instrument]
              properties:
                instrument:
                  type: string
                  example: "NQ1"
      responses:
        '200':
          description: Mag 7 analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mag7Analysis'
        '401':
          $ref: '#/components/responses/Unauthorized'
...